PG-33 (#19)

+++ backend/src/main/java/es/upm/backend/infrastructure/repository/impl/ClienteRepositoryImpl.java
@@ -1,5 +1,6 @@
 package es.upm.backend.infrastructure.repository.impl;
 
+import es.upm.backend.application.exception.ClienteNotFoundException;
 import es.upm.backend.domain.entities.Cliente;
 import es.upm.backend.domain.repository.ClienteRepository;
 import es.upm.backend.infrastructure.repository.jpa.ClienteJpaRepository;
@@ -13,10 +14,9 @@ public class ClienteRepositoryImpl implements ClienteRepository {
 
     private final ClienteJpaRepository clienteJpaRepository;
 
-    public ClienteRepositoryImpl(ClienteJpaRepository clienteJpaRepository){
+    public ClienteRepositoryImpl(ClienteJpaRepository clienteJpaRepository) {
         this.clienteJpaRepository = clienteJpaRepository;
     }
-
 
     @Override
     public List<Cliente> findAll() {
@@ -28,9 +28,28 @@ public class ClienteRepositoryImpl implements ClienteRepository {
         return clienteJpaRepository.save(cliente);
     }
 
-    @Override
+        @Override
     public void delete(Long idCliente) {
         Optional<Cliente> clienteAEliminar = clienteJpaRepository.findById(idCliente);
+        if (clienteAEliminar.isEmpty()) {
+            throw new ClienteNotFoundException(idCliente);
+        }
         clienteJpaRepository.delete(clienteAEliminar.get());
     }
+
+    @Override
+    public boolean existsByEmail(String email) {
+        return clienteJpaRepository.existsByEmail(email);
+    }
+
+    @Override
+    public Cliente save(Cliente cliente) {
+        return clienteJpaRepository.save(cliente); 
+    }
+
+    @Override
+    public boolean existsById(Long idCliente) {
+        return clienteJpaRepository.existsById(idCliente);
+    }
+
 }
